move miners around
